+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Set permissions for custom metadata types
 - Set permissions for custom permissions
+- Set permissions for visualforce pages
 - Internal structure changes
 - Bug fixes